Handled warnings

+++ src/f_different_powers.c
@@ -11,7 +11,7 @@ static void _different_powers_evaluate(coco_problem_t *self, const double *x,
 
   assert(self->number_of_objectives == 1);
   for (i = 0; i < self->number_of_variables; ++i) {
-    double exponent = 2.0 + (4.0 * i) / (self->number_of_variables - 1.0);
+    double exponent = 2.0 + (4.0 * (double)(long)i) / ((double)(long)self->number_of_variables - 1.0);
     sum += pow(fabs(x[i]), exponent);
   }
   y[0] = sqrt(sum);
@@ -23,11 +23,11 @@ different_powers_problem(const size_t number_of_variables) {
   coco_problem_t *problem = coco_allocate_problem(number_of_variables, 1, 0);
   problem->problem_name = coco_strdup("different powers function");
   /* Construct a meaningful problem id */
-  problem_id_length = snprintf(NULL, 0, "%s_%02i", "different powers",
-                               (int)number_of_variables);
+  problem_id_length = (size_t)snprintf(NULL, 0, "%s_%02lu", "different powers",
+                               (long)number_of_variables);
   problem->problem_id = coco_allocate_memory(problem_id_length + 1);
-  snprintf(problem->problem_id, problem_id_length + 1, "%s_%02d",
-           "different powers", (int)number_of_variables);
+  snprintf(problem->problem_id, problem_id_length + 1, "%s_%02lu",
+           "different powers", (long)number_of_variables);
 
   problem->number_of_variables = number_of_variables;
   problem->number_of_objectives = 1;
